Add printRawMatrix func move preset graph->newfile
